refactor involving folder organization and removing hashtag files
