+++ src/utils/paths.py
@@ -24,25 +24,18 @@ PATH_ORIGINAL_USERS = PATH_BASE / PATHS['data']['original_users']
 PATH_ORIGINAL_USERS_IDS = PATH_BASE / PATHS['data']['original_users_ids']
 
 # Data - Raw downloads
-PATH_HASHTAGS_RAW = PATH_BASE / PATHS['data']['hashtags_raw']
 PATH_TWEETS_RAW = PATH_BASE / PATHS['data']['tweets_raw']
 PATH_MENTIONS_RAW = PATH_BASE / PATHS['data']['mentions_raw']
 PATH_LOG = PATH_BASE / PATHS['data']['log']
 PATH_FOLDER_DATASET = PATH_BASE / PATHS['data']['folder_dataset']
 PATH_FOLDER_RAW = PATH_BASE / PATHS['data']['folder_raw']
 
-
 # Data - Raw downloads current
 def include_month_prefix(original_str):
     original_path = Path(original_str)
-
     month_prefix = time.strftime("%Y-%m")
-
     return str(Path(original_path.parent) / f"{month_prefix}_{original_path.name}")
-
-
 PATH_USERS_RAW_CURRENT = PATH_BASE / include_month_prefix(PATHS['data']['users_raw'])
-PATH_HASHTAGS_RAW_CURRENT = PATH_BASE / include_month_prefix(PATHS['data']['hashtags_raw'])
 PATH_TWEETS_RAW_CURRENT = PATH_BASE / include_month_prefix(PATHS['data']['tweets_raw'])
 PATH_MENTIONS_RAW_CURRENT = PATH_BASE / include_month_prefix(PATHS['data']['mentions_raw'])
 PATH_LOG_CURRENT = PATH_BASE / include_month_prefix(PATHS['data']['log'])
